# contracts/openapi.json
{
  "components": {
    "schemas": {
      "Business": {
        "properties": {
          "enterprise_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enterprise Id"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "receiving_processes": {
            "items": {
              "type": "string"
            },
            "title": "Receiving Processes",
            "type": "array"
          },
          "roles": {
            "items": {
              "enum": [
                "buyer",
                "seller"
              ],
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "roles",
          "receiving_processes",
          "location"
        ],
        "title": "Business",
        "type": "object"
      },
      "Costs": {
        "properties": {
          "buyer_total_paise": {
            "description": "INR paise",
            "minimum": 0.0,
            "title": "Buyer Total Paise",
            "type": "integer"
          },
          "freight_paise": {
            "description": "INR paise",
            "minimum": 0.0,
            "title": "Freight Paise",
            "type": "integer"
          },
          "material_paise": {
            "description": "INR paise",
            "minimum": 0.0,
            "title": "Material Paise",
            "type": "integer"
          },
          "seller_receives_paise": {
            "description": "INR paise",
            "minimum": 0.0,
            "title": "Seller Receives Paise",
            "type": "integer"
          }
        },
        "required": [
          "material_paise",
          "freight_paise",
          "buyer_total_paise",
          "seller_receives_paise"
        ],
        "title": "Costs",
        "type": "object"
      },
...
      "ExcludedListing": {
        "properties": {
          "listing_id": {
            "title": "Listing Id",
            "type": "string"
          },
          "reason_codes": {
            "items": {
              "type": "string"
            },
            "title": "Reason Codes",
            "type": "array"
          }
        },
        "required": [
          "listing_id",
          "reason_codes"
        ],
        "title": "ExcludedListing",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Listing": {
        "description": "Public listing. Deliberately has no seller floor field.",
        "properties": {
          "asking_price_paise_per_tonne": {
            "description": "INR paise per tonne",
            "minimum": 0.0,
            "title": "Asking Price Paise Per Tonne",
            "type": "integer"
          },
          "available_quantity_kg": {
            "description": "kilograms",
            "minimum": 0.0,
            "title": "Available Quantity Kg",
            "type": "integer"
          },
          "contamination_notes": {
            "title": "Contamination Notes",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "material_id": {
            "title": "Material Id",
            "type": "string"
          },
          "moisture_pct": {
            "title": "Moisture Pct",
            "type": "number"
          },
          "pickup_window": {
            "$ref": "#/components/schemas/Window"
          },
          "seller": {
            "$ref": "#/components/schemas/Business"
          },
          "status": {
            "enum": [
              "open",
              "closed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "seller",
          "material_id",
          "available_quantity_kg",
          "asking_price_paise_per_tonne",
          "moisture_pct",
          "contamination_notes",
          "pickup_window",
          "location",
          "status"
        ],
        "title": "Listing",
        "type": "object"
      },
...
      "Location": {
        "properties": {
          "district": {
            "title": "District",
            "type": "string"
          },
          "lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "lon": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lon"
          },
          "precision": {
            "default": "district",
            "enum": [
              "district",
              "confirmed_site"
            ],
            "title": "Precision",
            "type": "string"
          }
        },
        "required": [
          "district"
        ],
        "title": "Location",
        "type": "object"
      },
      "MatchCandidate": {
        "properties": {
          "initial_best_cost": {
            "$ref": "#/components/schemas/Costs"
          },
          "listing": {
            "$ref": "#/components/schemas/Listing"
          },
          "pathway_use": {
            "title": "Pathway Use",
            "type": "string"
          },
          "transport_options": {
            "items": {
              "$ref": "#/components/schemas/TransportOption"
            },
            "title": "Transport Options",
            "type": "array"
          }
        },
        "required": [
          "listing",
          "pathway_use",
          "transport_options",
          "initial_best_cost"
        ],
        "title": "MatchCandidate",
        "type": "object"
      },
      "MatchesResponse": {
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/MatchCandidate"
            },
            "title": "Candidates",
            "type": "array"
          },
          "excluded": {
            "items": {
              "$ref": "#/components/schemas/ExcludedListing"
            },
            "title": "Excluded",
            "type": "array"
          },
          "missing_transport_listing_ids": {
            "items": {
              "type": "string"
            },
            "title": "Missing Transport Listing Ids",
            "type": "array"
          },
          "requirement_id": {
            "title": "Requirement Id",
            "type": "string"
          }
        },
        "required": [
          "requirement_id",
          "candidates",
          "excluded",
          "missing_transport_listing_ids"
        ],
        "title": "MatchesResponse",
        "type": "object"
      },
...
      "OwnerRequirement": {
        "properties": {
          "buyer": {
            "$ref": "#/components/schemas/Business"
          },
          "buyer_max_total_paise": {
            "description": "INR paise",
            "minimum": 0.0,
            "title": "Buyer Max Total Paise",
            "type": "integer"
          },
          "delivery_window": {
            "$ref": "#/components/schemas/Window"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "location": {
            "$ref": "#/components/schemas/Location"
          },
          "material_id": {
            "title": "Material Id",
            "type": "string"
          },
          "max_moisture_pct": {
            "title": "Max Moisture Pct",
            "type": "number"
          },
          "quantity_kg": {
            "description": "kilograms",
            "minimum": 0.0,
            "title": "Quantity Kg",
            "type": "integer"
          },
          "receiving_process_id": {
            "title": "Receiving Process Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "open",
              "fulfilled",
              "closed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "buyer",
          "material_id",
          "receiving_process_id",
          "quantity_kg",
          "max_moisture_pct",
          "delivery_window",
          "location",
          "status",
          "buyer_max_total_paise"
        ],
        "title": "OwnerRequirement",
        "type": "object"
      },
...
      "Page_OwnerRequirement_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OwnerRequirement"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[OwnerRequirement]",
        "type": "object"
      },
...
      "RequirementCreate": {
        "properties": {
          "buyer_max_total_paise": {
            "description": "INR paise",
            "minimum": 0.0,
            "title": "Buyer Max Total Paise",
            "type": "integer"
          },
          "delivery_window": {
            "$ref": "#/components/schemas/Window"
          },
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Location"
              },
              {
                "type": "null"
              }
            ]
          },
          "material_id": {
            "title": "Material Id",
            "type": "string"
          },
          "max_moisture_pct": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Max Moisture Pct",
            "type": "number"
          },
          "quantity_kg": {
            "description": "kilograms",
            "exclusiveMinimum": 0.0,
            "title": "Quantity Kg",
            "type": "integer"
          },
          "receiving_process_id": {
            "title": "Receiving Process Id",
            "type": "string"
          }
        },
        "required": [
          "material_id",
          "receiving_process_id",
          "quantity_kg",
          "max_moisture_pct",
          "delivery_window",
          "buyer_max_total_paise"
        ],
        "title": "RequirementCreate",
        "type": "object"
      },
      "TransportOption": {
        "properties": {
          "capacity_kg": {
            "description": "kilograms",
            "minimum": 0.0,
            "title": "Capacity Kg",
            "type": "integer"
          },
          "delivery_at": {
            "format": "date-time",
            "title": "Delivery At",
            "type": "string"
          },
          "distance_basis": {
            "default": "unknown",
            "enum": [
              "road",
              "district_straight_line",
              "unknown"
            ],
            "title": "Distance Basis",
            "type": "string"
          },
          "distance_m": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance M"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "freight_paise": {
            "description": "INR paise",
            "minimum": 0.0,
            "title": "Freight Paise",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "listing_id": {
            "title": "Listing Id",
            "type": "string"
          },
          "pickup_at": {
            "format": "date-time",
            "title": "Pickup At",
            "type": "string"
          },
          "requirement_id": {
            "title": "Requirement Id",
            "type": "string"
          },
          "source": {
            "enum": [
              "entered_quote",
              "configured_estimate"
            ],
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "id",
          "listing_id",
          "requirement_id",
          "label",
          "freight_paise",
          "capacity_kg",
          "pickup_at",
          "delivery_at",
          "expires_at",
          "source"
        ],
        "title": "TransportOption",
        "type": "object"
      },
      "TransportOptionCreate": {
        "properties": {
          "capacity_kg": {
            "exclusiveMinimum": 0.0,
            "title": "Capacity Kg",
            "type": "integer"
          },
          "delivery_at": {
            "format": "date-time",
            "title": "Delivery At",
            "type": "string"
          },
          "distance_basis": {
            "default": "unknown",
            "enum": [
              "road",
              "district_straight_line",
              "unknown"
            ],
            "title": "Distance Basis",
            "type": "string"
          },
          "distance_m": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance M"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "freight_paise": {
            "description": "INR paise",
            "minimum": 0.0,
            "title": "Freight Paise",
            "type": "integer"
          },
          "label": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Label",
            "type": "string"
          },
          "listing_id": {
            "title": "Listing Id",
            "type": "string"
          },
          "pickup_at": {
            "format": "date-time",
            "title": "Pickup At",
            "type": "string"
          },
          "requirement_id": {
            "title": "Requirement Id",
            "type": "string"
          },
          "source": {
            "enum": [
              "entered_quote",
              "configured_estimate"
            ],
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "listing_id",
          "requirement_id",
          "label",
          "freight_paise",
          "capacity_kg",
          "pickup_at",
          "delivery_at",
          "expires_at",
          "source"
        ],
        "title": "TransportOptionCreate",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "Window": {
        "properties": {
          "end": {
            "format": "date-time",
            "title": "End",
            "type": "string"
          },
          "start": {
            "format": "date-time",
            "title": "Start",
            "type": "string"
          }
        },
        "required": [
          "start",
          "end"
        ],
        "title": "Window",
        "type": "object"
      }
    },
    "securitySchemes": {
      "AccessToken": {
        "description": "AUTH_MODE=dev: paste `dev:<email>` (for example `dev:buyer1@demo.bitnbuild.local`). AUTH_MODE=supabase: paste a Supabase access token.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "B2B exchange for secondary industrial materials. Prices are entered by participants; freight charges are entered or configured estimates. No market price feed, benchmark, or carrier booking is involved.",
    "title": "Karnataka Industrial Byproduct Exchange",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/me/requirements": {
      "get": {
        "operationId": "my_requirements_api_v1_me_requirements_get",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 100,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_OwnerRequirement_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessToken": []
          }
        ],
        "summary": "My Requirements",
        "tags": [
          "requirements"
        ]
      }
    },
...
    "/api/v1/requirements": {
      "post": {
        "operationId": "create_requirement_api_v1_requirements_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequirementCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnerRequirement"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessToken": []
          }
        ],
        "summary": "Create Requirement",
        "tags": [
          "requirements"
        ]
      }
    },
    "/api/v1/requirements/{requirement_id}": {
      "get": {
        "operationId": "get_requirement_api_v1_requirements__requirement_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "requirement_id",
            "required": true,
            "schema": {
              "title": "Requirement Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OwnerRequirement"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessToken": []
          }
        ],
        "summary": "Get Requirement",
        "tags": [
          "requirements"
        ]
      }
    },
    "/api/v1/requirements/{requirement_id}/matches": {
      "get": {
        "operationId": "get_matches_api_v1_requirements__requirement_id__matches_get",
        "parameters": [
          {
            "in": "path",
            "name": "requirement_id",
            "required": true,
            "schema": {
              "title": "Requirement Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessToken": []
          }
        ],
        "summary": "Get Matches",
        "tags": [
          "requirements"
        ]
      }
    },
    "/api/v1/transport-options": {
      "post": {
        "operationId": "create_transport_option_api_v1_transport_options_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransportOptionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransportOption"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "AccessToken": []
          }
        ],
        "summary": "Create Transport Option",
        "tags": [
          "requirements"
        ]
      }
    }
  }
}